+++ CHANGES.md
@@ -1,6 +1,18 @@
 stormpath-sdk-ruby Changelog
 ====================
 
+Version 0.4.0
+-------------
+
+Released on October 16, 2012
+
+- The Stormpath::Resource::GroupMembership class now extends the Stormpath::Resource::Resource class. It is no longer possible to call 'save' on an instance of this class.
+- The 'create' method of the Stormpath::Resource::GroupMembership class is now a class method and receives an instance of Stormpath::DataStore::DataStore.
+- The 'add_group' method implementation of the Stormpath::Resource::Account class was updated to reflect the previously mentioned changes.
+- The 'add_account' method implementation of the Stormpath::Resource::Group class was updated to reflect the previously mentioned changes.
+- The 'set_account' and 'set_group' methods were removed from the Stormpath::Resource::GroupMembership class.
+
+
 Version 0.3.0
 -------------